src/requirements.txt
aiosqlite==0.18.0
python-dotenv==1.0.0
python-telegram-bot[job-queue]==20.1
